Fixes compiler warning.
git-svn-id: http://svn.mythtv.org/svn/trunk@7691 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/videoout_ivtv.cpp
@@ -515,7 +515,7 @@ int VideoOutputIvtv::GetFirmwareFramesPlayed(void)
 int VideoOutputIvtv::GetFramesPlayed(void)
 {
     int frame = GetFirmwareFramesPlayed();
-    int f = internal_offset + (frame - frame_at_speed_change) * last_speed;
+    float f = internal_offset + (frame - frame_at_speed_change) * last_speed;
     return (int)round(f);
 }
